% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{decision_boundary}
\alias{decision_boundary}
\title{Decision Boundary to obtain the critical region R}
\usage{
decision_boundary(alpha, sigma, alternative = c("less", "greater"))
}
\arguments{
\item{alpha}{Two sided type I error.}

\item{sigma}{Estimated sigma for the future outcome \eqn{T_f}. In the case of log hazard ratio it can be considered as \eqn{\sqrt(4/d)}, where \eqn{d} are the number of events.}

\item{alternative}{Specify alternative (left tailed or right tailed)}
}
\value{
critical value
}
\description{
Calculates the decision boundary for a 1 sample design. This is the
critical value at which the decision function will change from 0
(failure) to 1 (success).
}
