% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{forestplot3}
\alias{forestplot3}
\title{Forest plot of the historical data. The K+1th study shows the predicted value of the future tretament effect.}
\usage{
forestplot3(out)
}
\arguments{
\item{out}{Output from JAGS from function "ppos_MAP"}
}
\value{
forest plot
}
\description{
Forest plot of the historical data. The K+1th study shows the predicted value of the future tretament effect.
}
