% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{input.data.prec.bhm}
\alias{input.data.prec.bhm}
\title{This function converts the within-study standard errors obtained from historical studies to variance-covariance or precision matrices required as an input
for jags model fit given the within-study correlartion coefficint estimate}
\usage{
input.data.prec.bhm(sigi, deltai, rho.hat)
}
\arguments{
\item{sigi}{a \eqn{K \times 1} within-study standard errors corresponding to primary outcome}

\item{deltai}{a \eqn{K \times 1} within-study standard errors corresponding to surrogate/secondary outcome}

\item{rho.hat}{user input of within-study correlation coefficient estimate}
}
\value{
a 3-d array of dimension \eqn{K \times m \times m} of precision matrices

a 3-d array of dimension \eqn{m \times m \times K} of variance-covaraince matrices
}
\description{
This function converts the within-study standard errors obtained from historical studies to variance-covariance or precision matrices required as an input
for jags model fit given the within-study correlartion coefficint estimate
}
