% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fn.R
\name{ppos}
\alias{ppos}
\title{Predictive probability of success PPoS}
\usage{
ppos(jags.output)
}
\arguments{
\item{jags.output}{Output from JAGS}
}
\value{
ppos value
}
\description{
Predictive probability of success PPoS
}
