% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imp_fn.R
\name{ppos_surrogate2}
\alias{ppos_surrogate2}
\title{Posterior analysis: Fits bivariate Bayesian hierarchical model to obtain the samples from posterior predictive distribution
of \eqn{T_f} to obtain the predictive probability of success (PPoS) given the critical region R from \code{decision_boundary}}
\usage{
ppos_surrogate2(
  Hy,
  Hy.se,
  rho.hat,
  tau.dist,
  tau.prior,
  alpha,
  X.cs,
  sigma.fs,
  sigma.cs,
  alternative,
  n.burnin = 4000,
  n.iter = 10000
)
}
\arguments{
\item{Hy}{Data set (historical) containing both final and surrogate endpoint from historical studies. A matrix of dimension K x 2 with first and second column containing final and surrogate endpoint treatment effect, respectively.}

\item{Hy.se}{Data set (historical) containing square of standard error of both final and surrogate endpoint from historical studies. A matrix of dimension K x 2 with first and second column containing final and surrogate endpoint variance (study specific), respectively.}

\item{rho.hat}{Estimate of study specific correlation coefficient.}

\item{tau.dist}{Type of prior distribution for between-study variance of both final and surrogate endpoint; supported priors are \code{HalfNormal} (default), \code{Uniform}, \code{InvGamma}, \code{LogNormal}, and \code{HalfCauchy}.}

\item{tau.prior}{Parameters of prior distribution for \code{tau}}

\item{alpha}{Two-sided alpha.}

\item{X.cs}{Observed treatment effect of surrogate endpoint in ongoing study.}

\item{sigma.fs}{Estimated variance of the treatment effect yet to be observed.}

\item{sigma.cs}{Variance of the treatment effect of surrogate endpoint observed in the ongoing trial.}

\item{alternative}{Specify alternative, left tailed = "less" and right tailed = "greater"}

\item{n.burnin}{Number of burn-in iterations in the MCMC.}

\item{n.iter}{Total number of iterations in the MCMC. Post burn-in = n.iter-n.burnin}
}
\value{
summary from JAGS output containing the estimates of all parameters in the Bayesian hierarchical model.
}
\description{
Posterior analysis: Fits bivariate Bayesian hierarchical model to obtain the samples from posterior predictive distribution
of \eqn{T_f} to obtain the predictive probability of success (PPoS) given the critical region R from \code{decision_boundary}
}
